requirements.txt
Flask
Werkzeug
line-bot-sdk
openai
python-dotenv
azure-core
azure-search-documents==11.4b9
google-generativeai
gunicorn
pypdf
azure-identity
azure-ai-formrecognizer
azure-storage-blob
pandas
streamlit
requests
pdfkit
beautifulsoup4